Call AppendEntries + debug put

+++ raft/coord.go
@@ -108,7 +108,7 @@ type JoinResponse struct {
 	ServerId uint8
 	Leader   bool
 	Peers    []ServerInfo
-	Term     uint8
+	Term     uint32
 }
 
 type HeadTailServerRequest struct {
@@ -149,8 +149,8 @@ type NotifyServerFailureAck struct {
 // the rightful leader
 type ServerLogState struct {
 	ServerId uint8
-	Term     uint8
-	LogIdx   int
+	Term     uint32
+	LogIdx   uint64
 }
 
 type ServerLogStateRequest struct {
